Added variable Money and connect it with
