+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Money.java
@@ -1,0 +1,22 @@
+package pl.com.bottega.ecommerce.sales.domain.offer;
+
+public class Money {
+    private int value;
+    private String currency;
+
+    public int getValue() {
+        return value;
+    }
+
+    public void setValue(int value) {
+        this.value = value;
+    }
+
+    public String getCurrency() {
+        return currency;
+    }
+
+    public void setCurrency(String currency) {
+        this.currency = currency;
+    }
+}
